toolkit/components/viewsource/test/test_428653.xul

<?xml-stylesheet href="chrome://global/skin/" type="text/css"?> 
<?xml-stylesheet
  href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="View Source Test (bug 428653)"
        xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
   src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
   src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
   
  <html:body/>

  <browser id="content" type="content-primary" name="content" src="about:blank" flex="1"
           disablehistory="true" context="viewSourceContextMenu"/>

  <script type="application/javascript">
  <![CDATA[
  /*
  Test that we can call the content browser's document.open() via
  wrappedJSObject (SafeJSObjectWrapper).
  https://bugzilla.mozilla.org/show_bug.cgi?id=428653
  */
  SimpleTest.waitForExplicitFinish();
  
  addLoadEvent(function testDocumentOpen() {
    var browser = document.getElementById("content");
    ok(browser, "got browser");
    var doc = browser.contentDocument.wrappedJSObject;
    ok(doc, "got content document");

    doc.open("text/html", "replace");

    var headingText = "This is a heading";
    doc.write("<h1 id='heading'>" + headingText + "</h1>");
    doc.close();

    ok(doc.getElementById("heading").textContent, headingText,
       "content document has content");
    
    SimpleTest.finish();
  });
  ]]>
  </script>
</window>
